requirements.txt
asttokens==2.4.1
astunparse==1.6.3
backcall==0.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
dependency-injector==4.41.0
Deprecated==1.2.14
executing==2.0.1
filelock==3.13.4
fsspec==2024.3.1
greenlet==3.0.3
huggingface-hub==0.22.2
idna==3.6
importlib_metadata==7.1.0
ipykernel==6.29.3
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jupyter_client==8.6.0
jupyter_core==5.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.0
progress==1.6
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
PyGithub==2.2.0
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
rich==13.7.1
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==2.6.1
six==1.16.0
SQLAlchemy==2.0.28
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.4.0
tokenizers==0.15.2
torch==2.2.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.39.3
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
wrapt==1.16.0
zipp==3.18.1
